+++ changelog.md
@@ -1,3 +1,9 @@
+0.2.0
+	* Modified HTTP class interface:
+		- Added defaults constructor parameter
+		- Changed return type of HTTP method calls to response
+		object instead of response body
+
 0.1.0
 
 	* Initial commit